# api/openapi.yaml
openapi: 3.0.3
info:
  title: Trellis Stateless Server
  version: 0.2.0
  description: |
    API for interacting with the Trellis State Machine Engine in a stateless manner.
servers:
  - url: http://localhost:{port}
    description: Local Dev Server
    variables:
      port:
        default: "8080"
        description: Server port
  - url: /
    description: Current Host (Relative)
paths:
...
  /navigate:
    post:
      summary: Transition to the next state based on input
      operationId: Navigate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NavigateRequest"
      responses:
        "200":
          description: New state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderResponse"
        "400":
          description: Invalid input
        "500":
          description: Internal server error
...
components:
  schemas:
    State:
      type: object
      required:
        - current_node_id
      properties:
        session_id:
          type: string
        current_node_id:
          type: string
          description: The identifier of the active node.
        memory:
          type: object
          description: Key-value store for session variables.
          additionalProperties: true
        history:
          type: array
          description: Trace of visited nodes.
          items:
            type: string
        terminated:
          type: boolean
          description: Indicates if the execution has reached a sink state.
        actions:
          type: array
          description: List of actions to be performed (e.g., render content).
          items:
            $ref: "#/components/schemas/ActionRequest"
    NavigateRequest:
      type: object
      required:
        - state
      properties:
        state:
          $ref: "#/components/schemas/State"
        input:
          description: The user input or tool result.
          oneOf:
            - type: string
            - $ref: "#/components/schemas/ToolResult"
    ToolResult:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Corresponds to the ToolCall ID.
        result:
          nullable: true
          description: The output of the tool execution.
        is_error:
          type: boolean
        error:
          type: string
    RenderResponse:
      type: object
      properties:
        state:
          $ref: "#/components/schemas/State"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/ActionRequest"
        terminal:
          type: boolean
          description: True if the current node has no outgoing transitions.
    ActionRequest:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: The type of action (e.g., RENDER_CONTENT, REQUEST_INPUT).
        payload:
          description: Dynamic payload depending on the action type.
          nullable: true
...